DataAccess method GetDbRecordCount changed to return int instead of double.

+++ BaeWynCouriersApp/DataAccess.cs
@@ -117,7 +117,7 @@ namespace BaeWynCouriersApp
             }
         }
 
-        public double GetDbRecordCount(string sqlstr)
+        public int GetDbRecordCount(string sqlstr)
         {
             try
             {
